# Use assocreduce in mapcompute instead of splatting