--- a/src/parallelism.jl
+++ b/src/parallelism.jl
@@ -9,7 +9,18 @@ function mapcompute(ctx, xs;
                     cache=false,
                     collect_results=identity,
                     map=map, kw...)
-    thunks = []
+
+    # Dagger does not have a way to just say "here is a bunch of Thunks, just compute them for me please"
+    # To make Dagger do this for us, we collect all Thunks in xs and ask Dagger to concatenate them into an array
+    # Once we have that array we just put the computed values back at the same place we found them
+    # Note that we rely on map visiting all elements in the same order each time it is called
+
+    # Creating all these intermediate arrys do have a little bit of overhead and I doubt this
+    # is the most efficient way of doing it. Hopefully if people end up here it is because they
+    # actually have heavy computations where the extra overhead is not significant
+
+    # Step 1: Collect  all thunks in xs
+    thunks = Thunk[]
     map(xs) do x
         if x isa Thunk
             if cache
@@ -20,8 +31,14 @@ function mapcompute(ctx, xs;
         x
     end
 
-    vals = collect_results(compute(ctx, delayed((xs...)->[xs...]; meta=true)(thunks...); kw...))
+    # Step 2: Ask Dagger to concatenate the results into vals
+    # We do assocreduce here mainly to prevent inference issues when xs is heterogenous
+    # Drawback vs e.g. splatting the whole array into a single delayed call to vcat is 
+    # that we end up creating a fair bit of intermediate arrays (about log2(length(thunks))).
+    vals = collect_results(compute(ctx, assocreduce(delayed(vcat; meta=true), thunks); kw...))
 
+    # Step 3: Put the computed results back at the same places we found them
+    # This is where we rely on map visiting all elements in the same order.
     i = 0
     map(xs) do x
         # Expression returns vals[i] or x
@@ -38,9 +55,11 @@ function mapexec(ctx, xs; cache=false, map=map)
     mapcompute(ctx, xs;
                map=map,
                cache=cache,
-               collect_results=xs -> asyncmap(d -> exec(ctx, d), xs))
+               collect_results=xs -> collect_chunks(ctx, xs))
 end
 
+collect_chunks(ctx, x) = [exec(ctx,x)]
+collect_chunks(ctx, xs::AbstractArray) = asyncmap(d -> exec(ctx,d), xs)
 
 """
     compute(tree::FileTree; cache=true)